+++ Source/Krypton Components/Krypton.Toolkit/Values/ToggleSwitchValues.cs
@@ -18,7 +18,10 @@ namespace Krypton.Toolkit
         #region Instance Fields
 
         private bool _enableEmbossEffect;
+        private bool _animateGradientEffect;
         private bool _enableKnobGradient;
+        private bool _onlyShowColorOnKnob;
+        private bool _showText;
         private float _gradientStartIntensity;
         private float _gradientEndIntensity;
         private LinearGradientMode _gradientDirection;
@@ -35,7 +38,10 @@ namespace Krypton.Toolkit
         public ToggleSwitchValues()
         {
             _enableEmbossEffect = false;
+            _animateGradientEffect = false;
             _enableKnobGradient = false;
+            _onlyShowColorOnKnob = true;
+            _showText = true;
             _gradientStartIntensity = 0.8f;
             _gradientEndIntensity = 0.6f;
             _gradientDirection = LinearGradientMode.ForwardDiagonal;
@@ -83,6 +89,22 @@ namespace Krypton.Toolkit
             }
         }
 
+        [Category("Appearance")]
+        [Description("Indicates whether the gradient effect should be animated.")]
+        [DefaultValue(false)]
+        public bool AnimateGradientEffect
+        {
+            get => _animateGradientEffect;
+            set
+            {
+                if (_animateGradientEffect != value)
+                {
+                    _animateGradientEffect = value;
+                    OnPropertyChanged(nameof(AnimateGradientEffect));
+                }
+            }
+        }
+
         /// <summary>Gets or sets a value indicating whether [enable knob gradient].</summary>
         /// <value><c>true</c> if [enable knob gradient]; otherwise, <c>false</c>.</value>
         [Category("Appearance")]
@@ -97,6 +119,40 @@ namespace Krypton.Toolkit
                 {
                     _enableKnobGradient = value;
                     OnPropertyChanged(nameof(EnableKnobGradient));
+                }
+            }
+        }
+
+        /// <summary>Gets or sets a value indicating whether [show color only on knob].</summary>
+        [Category("Appearance")]
+        [Description("Indicates whether the color should be only shown on the knob.")]
+        [DefaultValue(true)]
+        public bool OnlyShowColorOnKnob
+        {
+            get => _onlyShowColorOnKnob;
+            set
+            {
+                if (_onlyShowColorOnKnob != value)
+                {
+                    _onlyShowColorOnKnob = value;
+                    OnPropertyChanged(nameof(OnlyShowColorOnKnob));
+                }
+            }
+        }
+
+        /// <summary>Gets or sets a value indicating whether [show text].</summary>
+        [Category("Appearance")]
+        [Description("Indicates whether the text should be shown.")]
+        [DefaultValue(true)]
+        public bool ShowText
+        {
+            get => _showText;
+            set
+            {
+                if (_showText != value)
+                {
+                    _showText = value;
+                    OnPropertyChanged(nameof(ShowText));
                 }
             }
         }
@@ -204,6 +260,11 @@ namespace Krypton.Toolkit
                     value = 1;
                 }
 
+                if (value > 130)
+                {
+                    value = 130;
+                }
+
                 if (_cornerRadius != value)
                 {
                     _cornerRadius = value;
@@ -248,7 +309,10 @@ namespace Krypton.Toolkit
         public void Reset()
         {
             EnableEmbossEffect = false;
+            AnimateGradientEffect = false;
             EnableKnobGradient = false;
+            OnlyShowColorOnKnob = true;
+            ShowText = true;
             GradientStartIntensity = 0.8f;
             GradientEndIntensity = 0.6f;
             GradientDirection = LinearGradientMode.ForwardDiagonal;